Mouse：Double => Short completed

+++ client/src/main/scala/com/neo/sk/gypsy/model/GridOnClient.scala
@@ -430,7 +430,7 @@ class GridOnClient(override val boundary: Point) extends Grid {
   def reStart={
     myId = ""
     frameCount = 0l
-    food = Map[Point, Int]()
+    food = Map[Point, Short]()
     foodPool = 300
     playerMap = Map.empty[String,Player]
     virusMap = Map.empty[Long,Virus]
